closes #374 - init commit, creation of adminlayouts component, integrate it into adminlessons page
